Add Spotify user authorization
server/routes/authorizatonRouter.js has modified code from
https://github.com/spotify/web-api-auth-examples/blob/master/authorization_code/app.js.
The Apache License Version 2.0 has been applied to that code, so a copy
of it has been added.

Authorization starts by navigating to the login endpoint of the Express server
(ie. localhost:3001/login).

After the user has been successfully authorized, a document is added to the
users collection in MongoDB, or an existing document with the same
Spotify ID is updated with the new access and refresh tokens.