Move Spotify user authorization to client
The user interaction will be through the client, which includes user
authorization. This means that the redirection to the Spotify
authorization page occurs when the user goes to the login route on the
client side. After successfully authorizing the user, they are
redirected back to the callback route on the client side. The search
params of the callback URL are then parsed and sent to the Express server
through a POST request to make more requests to Spotify to get the tokens
and user information.