Send array of custom saved albums
The array of album IDs each user document has now includes the time
when it was added to their library. The album IDs in that array are
then used to query for all of their respective Album documents from
the database. The array of documents returned is processed to add
the time each album was added to the user's library, and then sent
to the client.