Fix how client and server use same route
Client and server shouldn't use the same route because it affects how
the browser will know what data to get from its cache. Previously, the
client and server had similar named routes, but now the server routes
are prefixed with 'api', as recommended here:
https://stackoverflow.com/a/56022718

This problem affected how when the user returns to this site, they will
only receive the JSON response from the API rather than the React
representation.