Refactor server and client
Some functions are added or converted in promises to make the code
more readable and work asynchronously.

Added /library/initialize endpoint which is used to get the initial or
default information for the client to render when the user first logs in.

Added /library/options endpoint which is used to get the possible
options for each sorting mode.

Added /library/update endpoint which is used to update the database
by retrieving and processing the current user's Spotify information.

Processing albums prior to adding to the database has changed
slightly. The duration in milliseconds is calculated, formatted, and
finally added to the document that will be added to the
database. Whether or not the album is explicit is also determined and
added to the document.

Modified /library endpoint so that it reads the query parameters of
the request and responds with the array of album IDs based on the
query parameters.

The client-side code has also changed based on the new
endpoints. Visually, the grid structure is missing for now, but it is
relatively the same as before.

Changed the variable name used in Controls component.

Once the Album component is mounted, it will request for the
information for the album ID prop. Unfortunately, it is very slow to
find the document for that album ID from the database. It may speed up
after making this application into a progressive web application and
caching data.