Add lazy loading
Album component only mounts when it is visible in the viewport. Only
when it mounts will it request for the album information.

The /library/album endpoint unnecessarily updated the user's library,
so it was removed. It must have been added accidentally when copying
the structure of an endpoint. This makes the endpoint respond
significantly faster than before.