Render Placeholder component during Album request
Before receiving the Album data, the Placeholder component is rendered
instead of actual Album component. This is so that when using a slower
network, there is feedback that the Album is loading.

Animation has been added to Placeholder. A transparent image is used
as a placeholder image instead. A background colour is added to the
image's div.