Redirect based on log in status
If the user is not logged in, they will be redirected to root route so
they can start log in process.

If the user is logged in and they navigate to either root, login, or
callback routes, they are redirected to the library route.

Implementation is a bit iffy though because it depends on a cookie
separate from the session cookie. If this cookie is somehow deleted
while the session cookie is not, the user will need to re-login
unnecessarily.