Log in console when cache needs to be updated
The user's library in the database is synced up with Spotify's library
when the Library component has mounted. The server sends the client
the updated information so that it can be compared with the data in
the cache. If there are any discrepancies, then that means that the
cache needs to be updated.