Add Reload button when service worker has updated
Whenever a new build is uploaded to the server, the service worker
itself is an updated version. This updated version is compared with
any existing service worker on the browser. If there is an updated
service worker, then a button will pop up to prompt the user to reload
the page and get the new changes.

The service worker and React app itself has been modified so that this
button appears at the correct situations.

Most of the new code is based on this example:
https://felixgerschau.com/2020/01/27/cra-pwa-update-notification.html
and
https://redfin.engineering/how-to-fix-the-refresh-button-when-using-service-workers-a8e27af6df68.