Add logout functionality
Provide visual feedback during logging in, authorization, and logging
out.

Added a menu that can be opened and closed. This menu contains the
buttons to download for offline use and logging out. Since the
download button is within this menu, Sort Plus will not check whether
or not the local copy can be updated until this menu is opened. It
might be a minor inconvenience.